POFIM-159: Fjerner metoder og restpunkt relatert til oppdatert forespørsel og tasker tilknyttet dette
